feat: update relayer addresses post oracle migration (#41)
- Add the new relayer addresses to mainnet now that the migration CGP
has executed
- Remove some addresses on staging that we no longer relay, since
Redstone is now providing the same pairs on Alfajores that they do on
mainnet, therefore the setup is 1 to 1. The only relayer that exists on
staging but not mainnet yet is for the _CELO/ETH_ feed, since that CGP
has not been proposed yet.

This is already deployed on both envs