feat: celo-sepolia network support (#43)
## Overview

This PR deploys a new relayer workspace called `sepolia`, in addition to
the existing `prod` and `staging` workspaces. This new workspace relays
the oracles on the `celo-sepolia` network.

## FAQ

### The workspaces are `prod`, `staging`, and `sepolia`, ugh?!

Yeah. The system right now has a 1:1 mapping between "workspace" and
"network", and there's no easy way to change that, i.e., deploy multiple
relayer functions in the "staging" workspace that point to different
networks.
Therefore, the easiest win was to create a whole new _temporary_
workspace called "sepolia", and once the alfajores network is officially
at end-of-life, we'll do a follow-up PR to change the network and
settings for the `staging` workspace to point to the sepolia network and
delete the `sepolia` workspace.

### `viem` updated? Did you check for supply chain injection?

Yup, to get the `celoSepolia` network, I had to do the update. But I
checked all packages manually 🗡️

### What's up with the `logger.ts` change with the type?

I honestly don't know, because that dependency didn't change, but
TypeScript was mad with the code as is, and the `LogEntry` was not the
expected type in the function. Not sure whether this will come back and
bite us in the ass, but I did what I had to do to make it happy.

---------

Co-authored-by: Philip Paetz <philip.paetz@me.com>
Co-authored-by: nvtaveras <4562733+nvtaveras@users.noreply.github.com>